// src/vs/base/common/marked/marked.d.ts
export type MarkedToken =
	| Tokens.Space
	| Tokens.Code
	| Tokens.Heading
	| Tokens.Table
	| Tokens.Hr
	| Tokens.Blockquote
	| Tokens.List
	| Tokens.ListItem
	| Tokens.Paragraph
	| Tokens.HTML
	| Tokens.Text
	| Tokens.Def
	| Tokens.Escape
	| Tokens.Tag
	| Tokens.Image
	| Tokens.Link
	| Tokens.Strong
	| Tokens.Em
	| Tokens.Codespan
	| Tokens.Br
	| Tokens.Del;
export type Token = MarkedToken | Tokens.Generic;
export declare namespace Tokens {
	interface Space {
		type: 'space';
		raw: string;
	}
	interface Code {
		type: 'code';
		raw: string;
		codeBlockStyle?: 'indented' | undefined;
		lang?: string | undefined;
		text: string;
		escaped?: boolean;
	}
	interface Heading {
		type: 'heading';
		raw: string;
		depth: number;
		text: string;
		tokens: Token[];
	}
	interface Table {
		type: 'table';
		raw: string;
		align: Array<'center' | 'left' | 'right' | null>;
		header: TableCell[];
		rows: TableCell[][];
	}
	interface TableRow {
		text: string;
	}
	interface TableCell {
		text: string;
		tokens: Token[];
		header: boolean;
		align: 'center' | 'left' | 'right' | null;
	}
	interface Hr {
		type: 'hr';
		raw: string;
	}
	interface Blockquote {
		type: 'blockquote';
		raw: string;
		text: string;
		tokens: Token[];
	}
	interface List {
		type: 'list';
		raw: string;
		ordered: boolean;
		start: number | '';
		loose: boolean;
		items: ListItem[];
	}
	interface ListItem {
		type: 'list_item';
		raw: string;
		task: boolean;
		checked?: boolean | undefined;
		loose: boolean;
		text: string;
		tokens: Token[];
	}
	interface Checkbox {
		checked: boolean;
	}
	interface Paragraph {
		type: 'paragraph';
		raw: string;
		pre?: boolean | undefined;
		text: string;
		tokens: Token[];
	}
	interface HTML {
		type: 'html';
		raw: string;
		pre: boolean;
		text: string;
		block: boolean;
	}
	interface Text {
		type: 'text';
		raw: string;
		text: string;
		tokens?: Token[];
	}
	interface Def {
		type: 'def';
		raw: string;
		tag: string;
		href: string;
		title: string;
	}
	interface Escape {
		type: 'escape';
		raw: string;
		text: string;
	}
	interface Tag {
		type: 'text' | 'html';
		raw: string;
		inLink: boolean;
		inRawBlock: boolean;
		text: string;
		block: boolean;
	}
	interface Link {
		type: 'link';
		raw: string;
		href: string;
		title?: string | null;
		text: string;
		tokens: Token[];
	}
	interface Image {
		type: 'image';
		raw: string;
		href: string;
		title: string | null;
		text: string;
	}
	interface Strong {
		type: 'strong';
		raw: string;
		text: string;
		tokens: Token[];
	}
	interface Em {
		type: 'em';
		raw: string;
		text: string;
		tokens: Token[];
	}
	interface Codespan {
		type: 'codespan';
		raw: string;
		text: string;
	}
	interface Br {
		type: 'br';
		raw: string;
	}
	interface Del {
		type: 'del';
		raw: string;
		text: string;
		tokens: Token[];
	}
	interface Generic {
		[index: string]: any;
		type: string;
		raw: string;
		tokens?: Token[] | undefined;
	}
}
export type Links = Record<string, Pick<Tokens.Link | Tokens.Image, 'href' | 'title'>>;
export type TokensList = Token[] & {
	links: Links;
};
/**
 * Renderer
 */
declare class MarkedRenderer {
	options: MarkedOptions;
	parser: MarkedParser;
	constructor(options?: MarkedOptions);
	space(token: Tokens.Space): string;
	code({ text, lang, escaped }: Tokens.Code): string;
	blockquote({ tokens }: Tokens.Blockquote): string;
	html({ text }: Tokens.HTML | Tokens.Tag): string;
	heading({ tokens, depth }: Tokens.Heading): string;
	hr(token: Tokens.Hr): string;
	list(token: Tokens.List): string;
	listitem(item: Tokens.ListItem): string;
	checkbox({ checked }: Tokens.Checkbox): string;
	paragraph({ tokens }: Tokens.Paragraph): string;
	table(token: Tokens.Table): string;
	tablerow({ text }: Tokens.TableRow): string;
	tablecell(token: Tokens.TableCell): string;
	/**
	 * span level renderer
	 */
	strong({ tokens }: Tokens.Strong): string;
	em({ tokens }: Tokens.Em): string;
	codespan({ text }: Tokens.Codespan): string;
	br(token: Tokens.Br): string;
	del({ tokens }: Tokens.Del): string;
	link({ href, title, tokens }: Tokens.Link): string;
	image({ href, title, text }: Tokens.Image): string;
	text(token: Tokens.Text | Tokens.Escape | Tokens.Tag): string;
}
/**
 * TextRenderer
 * returns only the textual part of the token
 */
declare class MarkedTextRenderer {
	strong({ text }: Tokens.Strong): string;
	em({ text }: Tokens.Em): string;
	codespan({ text }: Tokens.Codespan): string;
	del({ text }: Tokens.Del): string;
	html({ text }: Tokens.HTML | Tokens.Tag): string;
	text({ text }: Tokens.Text | Tokens.Escape | Tokens.Tag): string;
	link({ text }: Tokens.Link): string;
	image({ text }: Tokens.Image): string;
	br(): string;
}
/**
 * Parsing & Compiling
 */
declare class MarkedParser {
	options: MarkedOptions;
	renderer: MarkedRenderer;
	textRenderer: MarkedTextRenderer;
	constructor(options?: MarkedOptions);
	/**
	 * Static Parse Method
	 */
	static parse(tokens: Token[], options?: MarkedOptions): string;
	/**
	 * Static Parse Inline Method
	 */
	static parseInline(tokens: Token[], options?: MarkedOptions): string;
	/**
	 * Parse Loop
	 */
	parse(tokens: Token[], top?: boolean): string;
	/**
	 * Parse Inline Tokens
	 */
	parseInline(tokens: Token[], renderer?: MarkedRenderer | MarkedTextRenderer): string;
}
// ...
export declare const block: {
	normal: {
		blockquote: RegExp;
		code: RegExp;
		def: RegExp;
		fences: RegExp;
		heading: RegExp;
		hr: RegExp;
		html: RegExp;
		lheading: RegExp;
		list: RegExp;
		newline: RegExp;
		paragraph: RegExp;
		table: RegExp;
		text: RegExp;
	};
	gfm: Record<
		| 'code'
		| 'blockquote'
		| 'hr'
		| 'html'
		| 'table'
		| 'text'
		| 'heading'
		| 'list'
		| 'paragraph'
		| 'def'
		| 'fences'
		| 'lheading'
		| 'newline',
		RegExp
	>;
	pedantic: Record<
		| 'code'
		| 'blockquote'
		| 'hr'
		| 'html'
		| 'table'
		| 'text'
		| 'heading'
		| 'list'
		| 'paragraph'
		| 'def'
		| 'fences'
		| 'lheading'
		| 'newline',
		RegExp
	>;
};
// ...
declare class MarkedTokenizer {
	options: MarkedOptions;
	rules: Rules;
	lexer: MarkedLexer;
	constructor(options?: MarkedOptions);
	space(src: string): Tokens.Space | undefined;
	code(src: string): Tokens.Code | undefined;
	fences(src: string): Tokens.Code | undefined;
	heading(src: string): Tokens.Heading | undefined;
	hr(src: string): Tokens.Hr | undefined;
	blockquote(src: string): Tokens.Blockquote | undefined;
	list(src: string): Tokens.List | undefined;
	html(src: string): Tokens.HTML | undefined;
	def(src: string): Tokens.Def | undefined;
	table(src: string): Tokens.Table | undefined;
	lheading(src: string): Tokens.Heading | undefined;
	paragraph(src: string): Tokens.Paragraph | undefined;
	text(src: string): Tokens.Text | undefined;
	escape(src: string): Tokens.Escape | undefined;
	tag(src: string): Tokens.Tag | undefined;
	link(src: string): Tokens.Link | Tokens.Image | undefined;
	reflink(src: string, links: Links): Tokens.Link | Tokens.Image | Tokens.Text | undefined;
	emStrong(src: string, maskedSrc: string, prevChar?: string): Tokens.Em | Tokens.Strong | undefined;
	codespan(src: string): Tokens.Codespan | undefined;
	br(src: string): Tokens.Br | undefined;
	del(src: string): Tokens.Del | undefined;
	autolink(src: string): Tokens.Link | undefined;
	url(src: string): Tokens.Link | undefined;
	inlineText(src: string): Tokens.Text | undefined;
}
// ...
export interface MarkedExtension {
	/**
	 * True will tell marked to await any walkTokens functions before parsing the tokens and returning an HTML string.
	 */
	async?: boolean;
	/**
	 * Enable GFM line breaks. This option requires the gfm option to be true.
	 */
	breaks?: boolean | undefined;
	/**
	 * Add tokenizers and renderers to marked
	 */
	extensions?: TokenizerAndRendererExtension[] | undefined | null;
	/**
	 * Enable GitHub flavored markdown.
	 */
	gfm?: boolean | undefined;
	/**
	 * Hooks are methods that hook into some part of marked.
	 * preprocess is called to process markdown before sending it to marked.
	 * processAllTokens is called with the TokensList before walkTokens.
	 * postprocess is called to process html after marked has finished parsing.
	 */
	hooks?: HooksObject | undefined | null;
	/**
	 * Conform to obscure parts of markdown.pl as much as possible. Don't fix any of the original markdown bugs or poor behavior.
	 */
	pedantic?: boolean | undefined;
	/**
	 * Type: object Default: new Renderer()
	 *
	 * An object containing functions to render tokens to HTML.
	 */
	renderer?: RendererObject | undefined | null;
	/**
	 * Shows an HTML error message when rendering fails.
	 */
	silent?: boolean | undefined;
	/**
	 * The tokenizer defines how to turn markdown text into tokens.
	 */
	tokenizer?: TokenizerObject | undefined | null;
	/**
	 * The walkTokens function gets called with every token.
	 * Child tokens are called before moving on to sibling tokens.
	 * Each token is passed by reference so updates are persisted when passed to the parser.
	 * The return value of the function is ignored.
	 */
	walkTokens?: ((token: Token) => void | Promise<void>) | undefined | null;
}
export interface MarkedOptions extends Omit<
	MarkedExtension,
	'hooks' | 'renderer' | 'tokenizer' | 'extensions' | 'walkTokens'
> {
	/**
	 * Hooks are methods that hook into some part of marked.
	 */
	hooks?: MarkedHooks | undefined | null;
	/**
	 * Type: object Default: new Renderer()
	 *
	 * An object containing functions to render tokens to HTML.
	 */
	renderer?: MarkedRenderer | undefined | null;
	/**
	 * The tokenizer defines how to turn markdown text into tokens.
	 */
	tokenizer?: MarkedTokenizer | undefined | null;
	/**
	 * Custom extensions
	 */
	extensions?: null | {
		renderers: {
			[name: string]: RendererExtensionFunction;
		};
		childTokens: {
			[name: string]: string[];
		};
		inline?: TokenizerExtensionFunction[];
		block?: TokenizerExtensionFunction[];
		startInline?: TokenizerStartFunction[];
		startBlock?: TokenizerStartFunction[];
	};
	/**
	 * walkTokens function returns array of values for Promise.all
	 */
	walkTokens?: null | ((token: Token) => void | Promise<void> | (void | Promise<void>)[]);
}
// ...
export type MaybePromise = void | Promise<void>;
// ...
export declare namespace marked {
	const options: (options: MarkedOptions) => typeof marked;
	const setOptions: (options: MarkedOptions) => typeof marked;
	const getDefaults: typeof _getDefaults;
	const defaults: MarkedOptions;
	const use: (...args: MarkedExtension[]) => typeof marked;
	const walkTokens: (
		tokens: Token[] | TokensList,
		callback: (token: Token) => MaybePromise | MaybePromise[]
	) => MaybePromise[];
	const parseInline: {
		(
			src: string,
			options: MarkedOptions & {
				async: true;
			}
		): Promise<string>;
		(
			src: string,
			options: MarkedOptions & {
				async: false;
			}
		): string;
		(src: string, options?: MarkedOptions | undefined | null): string | Promise<string>;
	};
	const Parser: typeof MarkedParser;
	const parser: typeof MarkedParser.parse;
	const Renderer: typeof MarkedRenderer;
	const TextRenderer: typeof MarkedTextRenderer;
	const Lexer: typeof MarkedLexer;
	const lexer: typeof MarkedLexer.lex;
	const Tokenizer: typeof MarkedTokenizer;
	const Hooks: typeof MarkedHooks;
	const parse: typeof marked;
}
export declare const options: (options: MarkedOptions) => typeof marked;
export declare const setOptions: (options: MarkedOptions) => typeof marked;
export declare const use: (...args: MarkedExtension[]) => typeof marked;
export declare const walkTokens: (
	tokens: Token[] | TokensList,
	callback: (token: Token) => MaybePromise | MaybePromise[]
) => MaybePromise[];
export declare const parseInline: {
	(
		src: string,
		options: MarkedOptions & {
			async: true;
		}
	): Promise<string>;
	(
		src: string,
		options: MarkedOptions & {
			async: false;
		}
	): string;
	(src: string, options?: MarkedOptions | undefined | null): string | Promise<string>;
};
export declare const parse: typeof marked;
export declare const parser: typeof MarkedParser.parse;
export declare const lexer: typeof MarkedLexer.lex;